ipacm: Fix vlan id parameter type for GetMuxbyVID function
Change vlan_id type in the GetMuxbyVID function declaration
to uint16_t so it matches the function defintion in WAN.cpp.

+++ ipacm/inc/IPACM_Wan.h
@@ -375,7 +375,7 @@ public:
 #ifdef FEATURE_VLAN_MPDN
 	static ipacm_ipv4_wan_iface ipv4_to_iface[IPA_MAX_NUM_SW_PDNS];
 	static ipacm_ipv6_wan_iface ipv6_to_iface[IPA_MAX_NUM_SW_PDNS];
-	static int GetMuxByVid(uint8_t vlan_id, uint8_t *mux_id, ipa_ip_type iptype);
+	static int GetMuxByVid(uint16_t vlan_id, uint8_t *mux_id, ipa_ip_type iptype);
 	static int GetMTUByVid( uint16_t *mtu, uint16_t vlan_id, ipa_ip_type iptype);
 	static bool is_xlat_by_vid(uint16_t vlan_id);
 #endif
